requirements.txt
Flask==3.0.1
Flask_Cors==4.0.0
Flask_RESTful==0.3.10
matplotlib==3.8.2
numpy==1.26.3
Pillow==10.0.1
Pillow==10.2.0
Requests==2.31.0
torch==2.1.2
torchtext==0.6.0
torchvision==0.16.2
